+++ src/dtos/detailDto.js
