+++ swagger/swagger.js
@@ -18,9 +18,9 @@ const doc = {
   schemes: ["https", "http"],
 };
 
-const outputFile = "./swagger-output.json";  // 같은 위치에 swagger-output.json을 만든다.
+const outputFile = "./swagger-output.json"; 
 const endpointsFiles = [
-  "../index.js"  // 라우터가 명시된 곳을 지정해준다.
+  "../index.js"  
 ];
 
 // swaggerAutogen을 호출하여 Swagger 문서를 생성
